Added changes for api historical data for chart

+++ app/src/main/java/com/example/cryptoanalytic/screens/cryptocurrencyDetails/api/response/ConvertedVolume.kt
@@ -3,5 +3,5 @@ package com.example.cryptoanalytic.screens.cryptocurrencyDetails.api.response
 data class ConvertedVolume(
     val btc: Double,
     val eth: Double,
-    val usd: Long
+    val usd: Double
 )
